feat(fabric): fix hardcoded port 8443 and introduce the ordererAddress field
This commit removes all occurrences of the hardcoded port and introduces a new field called ordererAddress. This new field will be used to refer to the orderer address, similar to how peerAddress is used for the peer address.

Changes Made:
- Removed all occurrences of port 8443 in the Fabric codebase.
- Introduced the new ordererAddress field for referring to the orderer address.

By making this change, it is no longer necessary to rely on a proxy check when determining the orderer or peer address. This improves flexibility and makes the codebase more maintainable.

fixes: #2273

Signed-off-by: saurabhkumarkardam <saurabh.kumar.kardam@accenture.com>